+++ src/test/adapter-aleph/test/examples/aleph_handler.clj
@@ -1,0 +1,62 @@
+(ns test.examples.aleph-handler
+  (:require
+    [test.examples.common :as common]
+    [test.examples.counter :as counters]
+    [test.examples.form :as form]
+    [reitit.ring :as rr]
+    [starfederation.datastar.clojure.adapter.aleph :as aleph]))
+
+
+;; -----------------------------------------------------------------------------
+;; Counters
+;; -----------------------------------------------------------------------------
+(def update-signal (counters/->update-signal aleph/->sse-response))
+
+
+(defn increment
+  ([req]
+   (update-signal req inc))
+  ([req respond _raise]
+   (respond (increment req))))
+
+
+(defn decrement
+  ([req]
+   (update-signal req dec))
+  ([req respond _raise]
+   (respond (decrement req))))
+
+
+(def counter-routes
+  ["/counters/"
+   ["" {:handler #'counters/counters}]
+   ["increment/:id" #'increment]
+   ["decrement/:id" #'decrement]])
+
+
+
+;; -----------------------------------------------------------------------------
+;; Form
+;; -----------------------------------------------------------------------------
+(def endpoint (form/->endpoint aleph/->sse-response))
+
+
+(def form-routes
+  ["/form"
+   ["" {:handler #'form/form}]
+   ["/endpoint" {:middleware [common/wrap-mpparams]
+                 :handler #'endpoint}]])
+
+
+
+
+(def router
+  (rr/router
+    [counter-routes
+     form-routes]))
+
+
+(def handler
+  (rr/ring-handler router
+                   common/default-handler
+                   {:middleware common/global-middleware}))