+++ src/test/adapter-ring-jetty/test/ring_jetty_test.clj
@@ -1,9 +1,11 @@
 (ns test.ring-jetty-test
   (:require
-    [test.common :as common]
-    [test.examples.ring-handler :as rh]
-    [lazytest.core :as lt :refer [defdescribe expect it]]
-    [ring.adapter.jetty :as jetty]
+    [test.common                                  :as common]
+    [test.persistent-connection                   :as pc]
+    [test.smoketests                              :as st]
+    [test.examples.ring-handler                   :as rh]
+    [lazytest.core                                :as lt :refer [defdescribe expect it]]
+    [ring.adapter.jetty                           :as jetty]
     [starfederation.datastar.clojure.adapter.ring :as jetty-gen])
   (:import
     [org.eclipse.jetty.server Server ServerConnector]))
@@ -15,75 +17,59 @@
   (.stop server))
 
 
-(defn jetty-server-port [jetty-server]
-  (let [connector (-> jetty-server
-                      Server/.getConnectors
-                      seq
-                      first)]
-    (.getLocalPort ^ServerConnector connector)))
-
-
-(def ring-jetty-basic-opts
+(def ring-jetty-opts
   {:start! jetty/run-jetty
    :stop!  stop-jetty!
-   :get-port jetty-server-port
    :join? false})
 
 
 ;; -----------------------------------------------------------------------------
 ;; Tests
 ;; -----------------------------------------------------------------------------
-(defdescribe  counters-test
+(defdescribe smoke-tests
   {:webdriver true
-   :context [(common/with-server-f rh/handler ring-jetty-basic-opts)]}
+   :context [(st/with-server-f (assoc ring-jetty-opts
+                                      :handler rh/handler))]}
   (it "manages signals"
     (doseq [[driver-type driver] common/drivers]
-      (let [res (common/run-counters! @driver)]
-        (expect (= res common/expected-counters) (str driver-type))))))
+      (let [res (st/run-counters! @driver st/*port*)]
+        (expect (= res st/expected-counters) (str driver-type)))))
+
+  (it "manages forms"
+    (doseq [[driver-type driver] common/drivers]
+      (let [res (st/run-forms! @driver st/*port*)]
+        (expect (= res st/expected-form-vals) (str driver-type))))))
 
 
 (defdescribe counters-async-test
   {:webdriver true
-   :context [(common/with-server-f rh/handler
-                                   (assoc ring-jetty-basic-opts :async? true))]}
+   :context [(st/with-server-f (assoc ring-jetty-opts
+                                      :handler rh/handler
+                                      :async? true))]}
   (it "manages signals"
     (doseq [[driver-type driver] common/drivers]
-      (let [res (common/run-counters! @driver)]
-        (expect (= res common/expected-counters) (str driver-type))))))
+      (let [res (st/run-counters! @driver st/*port*)]
+        (expect (= res st/expected-counters) (str driver-type)))))
 
-;; -----------------------------------------------------------------------------
-(defdescribe form-test
-  {:webdriver true
-   :context [(common/with-server-f rh/handler ring-jetty-basic-opts)]}
   (it "manages forms"
     (doseq [[driver-type driver] common/drivers]
-      (let [res (common/run-form-test! @driver)]
-        (expect (= res common/expected-form-vals) (str driver-type))))))
-
-
-(defdescribe form-test-async
-  {:webdriver true
-   :context [(common/with-server-f rh/handler
-                                   (assoc ring-jetty-basic-opts :async? true))]}
-  (it "manages forms"
-    (doseq [[driver-type driver] common/drivers]
-      (let [res (common/run-form-test! @driver)]
-        (expect (= res common/expected-form-vals) (str driver-type))))))
+      (let [res (st/run-forms! @driver st/*port*)]
+        (expect (= res st/expected-form-vals) (str driver-type))))))
 
 
 ;; -----------------------------------------------------------------------------
 (defdescribe persistent-sse-test
-  "Testing persistent connection, events are sent from outide the ring handler."
-  {:context [(common/persistent-sse-f jetty-gen/->sse-response
-                                      (assoc ring-jetty-basic-opts
-                                             :async? true))]}
+  "Testing persistent connection, events are sent from outside the ring handler."
   (it "handles persistent connections"
-    (let [res (common/run-persistent-sse-test!)]
+    (let [res (pc/run-test (assoc ring-jetty-opts
+                                  :->sse-response jetty-gen/->sse-response
+                                  :async? true))]
       (expect (map? res))
-      (common/p-sse-status-ok? res)
-      (common/p-sse-http1-headers-ok? res)
-      (common/p-sse-body-ok? res))))
+      (pc/sse-status-ok? res)
+      (pc/sse-http1-headers-ok? res)
+      (pc/sse-body-ok? res))))
 
 
-
-
+(comment
+  (require '[lazytest.repl :as ltr])
+  (ltr/run-tests '[test.ring-jetty-test]))