+++ src/test/adapter-ring/starfederation/datastar/clojure/adapter/ring/impl_test.clj
@@ -48,6 +48,5 @@
 
 
 (comment
-  (user/reload!)
   (require '[lazytest.repl :as ltr])
   (ltr/run-test-var #'simple-test))